create the react vite app

+++ backend/controllers/productController.js
@@ -51,6 +51,9 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ success: false, message: "Invalid ID" });
+  }
   try {
     const deletedProduct = await Product.findByIdAndDelete(id);
     if (!deletedProduct) {
